+++ src/components/index.js
@@ -9,7 +9,7 @@ export { default as Email } from './email';
 export { default as Footer } from './footer';
 export { default as Hero } from './sections/Hero';
 export { default as About } from './sections/About';
-export { default as Jobs } from './sections/jobs';
-export { default as Featured } from './sections/featured';
+export { default as Jobs } from './sections/Jobs';
+export { default as Featured } from './sections/Featured';
 export { default as Projects } from './sections/projects';
 export { default as Contact } from './sections/contact';
