+++ src/components/sections/Featured.jsx
@@ -348,7 +348,7 @@ const Featured = () => {
   return (
     <section id="projects">
       <h2 className="numbered-heading" ref={revealTitle}>
-        Some Things I’ve Built
+        Some Things I&apos;ve Built
       </h2>
 
       <StyledProjectsGrid>
